Added alchemy menu comp and cleaned up prev commit
