+++ client/src/pages/Stage1.jsx
@@ -8,6 +8,7 @@ import GameNav from "../components/GameNav";
 import Options from "../components/Options";
 import Player from "../components/Player";
 import Menu from "../components/Menu";
+import AlchemyMenu from "../components/AlchemyMenu";
 import PlayerCreature from "../components/PlayerCreature";
 import EnemyCreature from "../components/EnemyCreature";
 import creatures from "../constants/creatures";
@@ -435,7 +436,7 @@ function Stage1() {
 
           <Player player={player} />
 
-          {/* menu and creatures wrapped in options status check */}
+          {/* menus and creatures wrapped in options status check */}
           {!optionsStatus ? (
             <>
               <Menu
@@ -454,6 +455,13 @@ function Stage1() {
                 setBattleUndecided={setBattleUndecided}
                 setSpawnAnimation={setSpawnAnimation}
                 loadDataAlchemy={loadDataAlchemy}
+                createPotion={createPotion}
+                consumePotion={consumePotion}
+              />
+
+              <AlchemyMenu
+                gameMenuStatus={gameMenuStatus}
+                setGameMenuStatus={setGameMenuStatus}
                 createPotion={createPotion}
                 consumePotion={consumePotion}
               />
